requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
dataclasses-json==0.6.7
distro==1.9.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langsmith==0.3.45
load-dotenv==0.1.0
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.1
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.0
openai==1.84.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.25.1
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.45.1
tenacity==9.1.2
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.3.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.1
zstandard==0.23.0
